+++ packages/at_auth/test/legacy_at_keys_test.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:at_auth/at_auth.dart';
+import 'package:at_auth/at_auth_io.dart';
 import 'package:at_commons/at_commons.dart';
 import 'package:test/test.dart';